fix(wechatcomapp): split long text messages into multiple parts
