+++ channel/wechatmp/passive_reply.py
@@ -11,6 +11,7 @@ from channel.wechatmp.common import *
 from channel.wechatmp.wechatmp_channel import WechatMPChannel
 from channel.wechatmp.wechatmp_message import WeChatMPMessage
 from common.log import logger
+from common.utils import split_string_by_utf8_length
 from config import conf
